refactor: moved hash function to util.go
